// src/app/models/medicamento.model.ts
export class MedicamentoModel {
  id: string;
  nombre: string;
  envase: number;
  presentacion: string;
  precio: number;
  disponibilidad: number;

  // constructor( id: string, nombre: string, envase: number, presentacion: string, precio: number, disponibilidad: number) {
  //   this.id = id;
  //   this.nombre = nombre;
  //   this.envase = envase;
  //   this.presentacion = presentacion;
  //   this.precio = precio;
  //   this.disponibilidad = disponibilidad;
  // }
}